+++ cli/col.c
@@ -3,6 +3,32 @@
 #include "col.h"
 #include "cli.h"
 
+static int show_columns(char *prj)
+{
+    // TODO: add support for project columns
+    struct column {
+        char mark;
+        char tag[10];
+        char desc[80];
+    } columns[] = {
+        {.tag = "uknw",.mark = '?',.desc = "unknown tag"},
+        {.tag = "curr",.mark = '*',.desc = "current task"},
+        {.tag = "prev",.mark = '^',.desc = "previous task"},
+        {.tag = "blog",.mark = '+',.desc = "backlog column"},
+
+// user define columns (gotta read 'em from config file
+        {.tag = "revw",.mark = '>',.desc = "review column"},
+        {.tag = "test",.mark = '$',.desc = "test column"},
+        {.tag = "lock",.mark = '!',.desc = "locked column"},
+        {.tag = "done",.mark = '-',.desc = "done column"},
+    };
+
+    for (int i = 0; i < sizeof(columns) / sizeof(columns[0]); ++i)
+        printf("%c %s \t\t%s\n", columns[i].mark, columns[i].tag,
+               columns[i].desc);
+    return 0;
+}
+
 // TODO: Find a good error message in case option fails.  */
 int tman_cli_col(int argc, char **argv, struct tman_context *ctx)
 {
@@ -36,8 +62,7 @@ int tman_cli_col(int argc, char **argv, struct tman_context *ctx)
     if (showhelp == 1)
         return help_usage("col");
     else if (showlist == 1) {
-        printf("show list of columns\n");
-        return TMAN_OK;
+        return show_columns("");
     } else if ((col = argv[optind++]) == NULL)
         return elog(1, "gotta specify column to move a task to");